refactor: move tools to sub dir
