+++ src/tools/index.ts
@@ -6,7 +6,7 @@ import { getApp } from "./kintone/app/get-app.js";
 import { getApps } from "./kintone/app/get-apps.js";
 import { getFormFields } from "./kintone/app/get-form-fields.js";
 import { getProcessManagement } from "./kintone/app/get-process-management.js";
-import { updateStatuses } from "./kintone/update-statuses.js";
+import { updateStatuses } from "./kintone/record/update-statuses.js";
 import type { Tool } from "./utils.js";
 
 export const tools: Array<Tool<any, any>> = [
