+++ src/tools/kintone/record/__tests__/update-statuses.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
 import { updateStatuses } from "../update-statuses.js";
 import { z } from "zod";
-import { mockExtra, mockKintoneConfig } from "../../../__tests__/utils.js";
+import { mockExtra, mockKintoneConfig } from "../../../../__tests__/utils.js";
 
 // Mock the KintoneRestAPIClient
 const mockUpdateRecordsStatus = vi.fn();
